Writing Creational:Builder design pattern.
This example how to instantiate a Car object building its parts.
The constructor of Car is not public, so only CarBuilder can construct a Car because it is a friend class.
The aim of Builder design pattern is to be able to instantiate complex objects in a safe way.
It is possible also to set a variable that allows to return the full constructed object only if the process of construction is done.