Adding Factory Design Pattern
The factory design pattern is used to handle the responsabilities of creating an object of classes to another class, the factory class.
With this is possible to hide the interface of the desired classes.
In this example, the client cannot call the constructor of Animal, Dog and Cat, the only way to obtain an object is to use the Factory class.