Adding Singleton design pattern
Singleton design pattern is one of the most hated design patterns, because it is normally associated with global variables.
Also it is very simple to use it in a wrong way.
The aim of the singleton design pattern is to instatiate a object only once, in other words, having just one copy of the object on the system.
It can be used in the right way in just a few rare cases. Somewhat can be used on loggin systems.
The best way to use it is by depedency injection, in other words, do not using it directly, instead using as a depedency for other classes.