Adding Adapter Design Pattern
The main aim of the Design Pattern is to adapt interfaces of Legacy classes that cannot be updated (ex libraries) to a interface
that the Client wants.
In this example I create a Adapter that calls Huawei producer, Tesla producer, and adapt the interface from OldCarProducer from build to produce.