Implementing the Bridge Design Pattern
The aim of the Bridge design pattern is to separate the interface from the implementation.
On this example we have the class CompanyTraveller that uses the method doTravel(), but each object does not know what type of implementation would be used.
So the Traveler does not care about the implementation of the travel method, because the implementation is defined by the class TravelMethod.
In other words, the Traveler and the TravelMethod can vary independently