Implementing Composite Design Pattern
The Composite Design Pattern allows to create composed of other objects, in a tree style.
It is easy to add new classes to this design, it only needs to inherit from the Composite class and implement the virtual operations.
With this approach it is possible to do operations that can run recursevely passing by all objects within the tree.
In this example There are sectors and employees, the sectors have a vector of the Composite class, which can be used to store more sectors and employees.
the method calculateValue sums the value from each object and return the amount of a subtree.