Implementing Decorator Design Pattern
The aim of Decorator Design Pattern is to be able to decorate an object, in other words, adding some properties that can enhance the object.
On this example we have Orc and Minotaur monsters, also two decorators: Fly and Boss.
A commum Orc can be created, but also it can be anhanced into a flying monster and/or a very strong boss.