Implementing Facade Design Pattern
The main aim of facade is to make an easy interface to be used for the client.
But the easy interface hides more complex methods that the Facade takes care of.
In this example the HumanFacade has the simple method feedDog(), but this method hides a more complex implementation.
To feed the dog it is necessary to create a dog, create a DogFood, create a Bow based on DogFood and use use the method eat() of Dog.