Implementing Flyweight Design Pattern
The aim of the Flyweight Design Pattern is to share data that does not need to be created again.
It can be done by multiple objects pointing to the same shared data.
The Pattern is used on this example on the management of a City that register Persons.
Instead of adding each first name and last name, it can add just a reference to a list with all names combined.
In this case, each reference only use 4 bytes of memory, instead of the whole string.