Implementing Proxy Design Pattern
The aim of the Proxy Design Pattern is to include new features on a already created class but without modifying its source code.
In this example we have the BigQuery class which loads the file even if it is not used.
The LazyQuery class is a Proxy for the BigQuery, which only loads the file when the method call is called.