Implementing Chain Of Responsability Design Pattern
The aim of this pattern is to create a ordered list to process a message.
After one object process one message, it is passed to the next one on the list.
The example in this code uses the Design Pattern to sum the damage on a group of monsters.
Then a set of heroes individually try to confront those monsters.