Iterator Design Pattern
The aim of the iterator design pattern is to visit elements sequentially in a complex structure like default arrays.
It is possible to use iterators to implement a order of visits in trees, for example.
This example uses the iterator from std::vector, which can visit all elements on a vector from the first position to the last,
and also in a reverse order.