Implementinf Null Object Design Pattern
The aim of the Null Object Design Pattern is to create compatibility with classes that needs objects that might be null.
The implementation of the pattern in this code shows the File Class that needs a Log object to be created, even if you do not want one.
To solve this problem the NullLog implements the inferface of the Log Class but, in fact, the object does nothing.