Implementing Template Design Pattern
The aim of the Design Pattern is to create basic interfaces that can be inherited to get the virtual.
It has the same goal as the Strategy pattern, but uses inheritance instead.
In this code the pattern is used to force the database subclasses to have methods to interact with the database.
They are not pure virtual because maybe some subclass won't need a specific method.